+++ src/media/audio/acoustic/mod.rs
@@ -1,0 +1,17 @@
+// devela::media::audio::acoustic
+//
+#![doc = crate::_tags!(audio wip)]
+#![doc = crate::_DOC_MEDIA_AUDIO_ACOUSTIC!()]
+#![doc = crate::_doc!(modules: crate::media::audio; acoustic)]
+#![doc = crate::_doc!(flat:"media")]
+#![doc = crate::_doc!(hr)]
+
+// mod spatial;
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     spatial::*,
+        // };
+    }
+}