+++ src/media/audio/mod.rs
@@ -1,31 +1,32 @@
 // devela::media::audio
 //
 #![doc = crate::_DOC_MEDIA_AUDIO!()] // public
-#![doc = crate::_doc!(modules: crate::media; audio)]
+#![doc = crate::_doc!(modules: crate::media; audio)] // acoustic, format, music, synth
 #![doc = crate::_doc!(flat:"media")]
 #![doc = crate::_doc!(hr)]
 //
 // safety
 #![cfg_attr(feature = "safe_audio", forbid(unsafe_code))]
 
-// mod codec; // …
-// mod drum_machine;
-mod layout; // AudioChannel, AudioChannels
-// mod music; //
-mod pcm; // Pulse-code modulation audio buffers and stream metadata.
-// mod spatial; // …
-// mod synth; // …
+// mod acoustic; // Acoustic spaces, sources, listeners, propagation, effects
+// mod effect;   // Signal filters, dynamics, delays, and transformations
+// mod format;    // Encoded audio representations and containers
+// mod instrument; // Instruments, sample maps, presets, and sound banks
+mod layout; // Channel roles, arrangements, and sample layouts
+// mod music;    // Theory, tuning, notation, harmony, performance data
+mod pcm; // PCM samples, buffers, specs, and stream metadata
+// mod rhythm;   // Pulses, meters, grooves, trigger patterns
+// mod synth;    // Oscillators, envelopes, voices, modulation
 
 crate::structural_mods! { // _mods
     _mods {
         pub use super::{
-            // codec::*,
-            // drum_machine::*,
+            // acoustic::_all::*,
+            // codec::_all::*,
             layout::*,
-            // music::*,
-            pcm::*,
-            // spatial::*,
-            // synth::*,
+            // music::_all::*,
+            pcm::_all::*,
+            // synth::_all::*,
         };
     }
 }