+++ src/media/audio/pcm/mod.rs
@@ -7,14 +7,12 @@
 mod tests;
 
 mod buffer; // PcmBuffer
-mod format; // PcmWav*
 mod spec; // PcmSpec, PcmSample
 
 crate::structural_mods! { // _mods
     _mods {
         pub use super::{
             buffer::*,
-            format::*,
             spec::*,
         };
     }