+++ src/media/motion/mod.rs
@@ -1,0 +1,27 @@
+// devela::media::motion
+//
+#![doc = crate::_DOC_MEDIA_MOTION!()] // public
+#![doc = crate::_doc!(modules: crate::media; motion)]
+#![doc = crate::_doc!(flat:"media")]
+#![doc = crate::_doc!(hr)]
+
+// mod animation;   // composition of timed changes
+// mod object;      // primary motion
+
+// mod camera;      // secondary/view motion
+// mod curve;       // easing, envelopes, splines
+
+// mod gesture;     // expressive motion primitives
+// mod rhythm;      // pulses, meters, cycles, trigger grids
+
+// mod time;        // normalized time, phase, progress
+// mod transform;   // translation, rotation, scale
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     object::*,
+        //     camera::*,
+        // };
+    }
+}